day 1 part 1
Read the file as strings, instead of parsing it as lisp with `read`,
then uses `parse-integer` to tell when we've moved to the next elf and
sum or maximize accordingly.

Had to cheat a little with `setf` to reset the per-elf sum to 0, but I
wasn't seeing any alternative. Maybe something with `iterate`
generators?